finished - tests pass

+++ lib/trip.rb
@@ -31,7 +31,7 @@ module RideShare
         driver_id = object[1].to_i
         rider_id = object[2].to_i
         date = object[3].to_s
-        rating = object[4].to_i
+        rating = object[4].to_f
         a_trip = RideShare::Trip.new(trip_id, driver_id, rider_id, date, rating)
         all_trips_array << a_trip
       end
